search customer get by text File
